Microservicios/services/gateway/requirements.txt
Flask==2.3.3
PyJWT==2.8.0
requests==2.31.0
redis==5.0.1
python-dotenv==1.0.0
